backRag/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "TránsitoBot API - Código Nacional de Tránsito de Colombia"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Core FastAPI dependencies
    "fastapi>=0.104.0,<0.110.0",
    "uvicorn[standard]>=0.24.0,<0.30.0",
    "pydantic>=2.4.0,<2.8.0",
    "pydantic-settings>=2.0.0,<3.0.0",
    "python-multipart>=0.0.6,<0.1.0",
    "python-dotenv>=1.0.0,<1.1.0",
    # ChromaDB and embeddings
    "chromadb>=0.4.0,<0.5.0",
    "sentence-transformers>=2.2.0,<2.8.0",
    # Document processing
    "python-docx>=0.8.11,<1.2.0",
    # LLM integration (Anthropic Claude)
    "anthropic>=0.7.0,<1.0.0",
    # OpenRouter integration
    "openai>=1.0.0,<2.0.0",
    # ML dependencies
    "numpy>=1.21.0,<1.27.0",
    "torch>=1.13.0,<2.2.0",
    "transformers>=4.21.0,<4.40.0",
    "accelerate>=0.20.0,<0.30.0",
    # Utilities
    "requests>=2.28.0,<3.0.0",
    "typing-extensions>=4.5.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0,<8.0.0",
    "pytest-asyncio>=0.21.0,<0.22.0",
    "httpx>=0.25.0,<0.26.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]